+++ packages/headless/src/case-assist.index.ts
@@ -24,6 +24,7 @@ export * from './features/case-input/case-input-actions-loader';
 export * from './features/case-field/case-field-actions-loader';
 export * from './features/document-suggestion/document-suggestion-actions-loader';
 export * from './features/case-assist/case-assist-analytics-actions-loader';
+export * from './features/analytics/generic-analytics-actions-loader';
 
 // Controllers
 export type {