+++ packages/headless/src/insight.index.ts
@@ -32,6 +32,7 @@ export * from './features/case-context/case-context-actions-loader';
 export * from './features/insight-search/insight-search-analytics-actions-loader';
 export * from './features/fields/fields-actions-loader';
 export * from './features/attached-results/attached-results-actions-loader';
+export * from './features/analytics/generic-analytics-actions-loader';
 
 // Controllers
 export type {
@@ -320,3 +321,6 @@ export type {
 } from './api/search/date/relative-date';
 
 export {deserializeRelativeDate} from './api/search/date/relative-date';
+
+export {getOrganizationEndpoints} from './api/platform-client';
+export type {PlatformEnvironment} from './utils/url-utils';